fixup! refactor: group webapp routes to allow opting out of header/footer layout
